# Conflicts:
#	src/main/java/de/uniks/stp24/component/game/SitesComponent.java
#	src/main/java/de/uniks/stp24/controllers/InGameController.java
#	src/main/java/de/uniks/stp24/service/game/EventService.java
#	src/main/java/de/uniks/stp24/service/game/ResourcesService.java
#	src/test/java/de/uniks/stp24/AppTest2.java
#	src/test/java/de/uniks/stp24/game/islandOverview/IslandOverviewTestComponent.java
#	src/test/java/de/uniks/stp24/game/islandOverview/IslandOverviewTestInitializer.java

+++ src/main/java/de/uniks/stp24/component/game/DistrictComponent.java
@@ -82,6 +82,6 @@ public class DistrictComponent extends VBox implements ReusableItemComponent<Sit
 
     public void setJobProgress(Job job) {
         this.jobProgressBox.setVisible(true);
-        this.jobTimeText.setText(String.format("%s/%s", job.progress(), job.total()));
+        this.jobTimeText.setText(String.format("%s/%s", job.progress(), (int) job.total()));
     }
 }
